+++ Source/SwiftLintBuiltInRules/Rules/RuleConfigurations/ModifierOrderConfiguration.swift
@@ -10,6 +10,7 @@ struct ModifierOrderConfiguration: SeverityBasedRuleConfiguration {
     @ConfigurationElement(key: "preferred_modifier_order")
     private(set) var preferredModifierOrder: [SwiftDeclarationAttributeKind.ModifierGroup] = [
         .override,
+        .isolation,
         .acl,
         .setterACL,
         .dynamic,